common_function::aggrs::aggr_wrapper

Function aggr_merge_func_name

pub fn aggr_merge_func_name(aggr_name: &str) -> String

Returns the name of the merge function for the given aggregate function name. The merge function is used to merge the states of the state functions. The merge function’s name is in the format `__<aggr_name>_merge