common_function::aggrs::approximate

Module hll

Two UDAFs are implemented for HyperLogLog:

hll: Accepts a string column and aggregates the values into a HyperLogLog state.

hll_merge: Accepts a binary column of states generated by hll and merges them into a single state.

The states can be then used to estimate the cardinality of the values in the column by hll_count UDF.

Structs

HllState

Constants

DEFAULT_PRECISION 🔒

HLL_MERGE_NAME

HLL_NAME

Type Aliases

HllStateType 🔒