common_function
Module handlers

Traits

FlowServiceHandler

This flow service handler is only use for flush flow for now.

ProcedureServiceHandler

A trait for handling procedure service requests in QueryEngine.

TableMutationHandler

A trait for handling table mutations in QueryEngine.

Type Aliases

FlowServiceHandlerRef

ProcedureServiceHandlerRef

TableMutationHandlerRef