common_function::handlers

Trait ProcedureServiceHandler

pub trait ProcedureServiceHandler: Send + Sync {
    // Required methods
    fn migrate_region<'life0, 'async_trait>(
        &'life0 self,
        request: MigrateRegionRequest,
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn reconcile<'life0, 'async_trait>(
        &'life0 self,
        request: ReconcileRequest,
    ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query_procedure_state<'life0, 'life1, 'async_trait>(
        &'life0 self,
        pid: &'life1 str,
    ) -> Pin<Box<dyn Future<Output = Result<ProcedureStateResponse>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn manage_region_follower<'life0, 'async_trait>(
        &'life0 self,
        request: ManageRegionFollowerRequest,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn catalog_manager(&self) -> &CatalogManagerRef;
}

A trait for handling procedure service requests in QueryEngine.

Required Methods

fn migrate_region<'life0, 'async_trait>( &'life0 self, request: MigrateRegionRequest, ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Migrate a region from source peer to target peer, returns the procedure id if success.

fn reconcile<'life0, 'async_trait>( &'life0 self, request: ReconcileRequest, ) -> Pin<Box<dyn Future<Output = Result<Option<String>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Reconcile a table, database or catalog, returns the procedure id if success.

fn query_procedure_state<'life0, 'life1, 'async_trait>( &'life0 self, pid: &'life1 str, ) -> Pin<Box<dyn Future<Output = Result<ProcedureStateResponse>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Query the procedure’ state by its id

fn manage_region_follower<'life0, 'async_trait>( &'life0 self, request: ManageRegionFollowerRequest, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Manage a region follower to a region.

fn catalog_manager(&self) -> &CatalogManagerRef

Get the catalog manager