common_function::scalars::json

Module json_get

Macros

json_get 🔒

Get the value from the JSONB by the given path and return it as specified type. If the path does not exist or the value is not the type specified, return NULL.

Structs

JsonGetBool

Get the value from the JSONB by the given path and return it as a boolean.

JsonGetFloat

Get the value from the JSONB by the given path and return it as a float.

JsonGetInt

Get the value from the JSONB by the given path and return it as an integer.

JsonGetObject 🔒

Get the object from JSON value by path.

JsonGetString

Get the value from the JSONB by the given path and return it as a string.

Functions

get_json_by_path 🔒