Enum common_function::scalars::matches::PatternAst

enum PatternAst {
    Literal {
        op: UnaryOp,
        pattern: String,
    },
    Binary {
        op: BinaryOp,
        children: Vec<PatternAst>,
    },
    Group {
        op: UnaryOp,
        child: Box<PatternAst>,
    },
}

Variants

Literal

A leaf node that matches a column with pattern

Fields

op: UnaryOp

pattern: String

Binary

Flattened binary chains

Fields

op: BinaryOp

children: Vec<PatternAst>

Group

A sub-tree enclosed by parenthesis

Fields

op: UnaryOp

child: Box<PatternAst>

Implementations

impl PatternAst

fn into_like_expr(self, column: &str) -> Expr

fn convert_literal(column: &str, pattern: &str) -> Expr

fn escape_pattern(pattern: &str) -> String

fn transform_ast(self) -> Result<Self>

Transform this AST with preset rules to make it correct.

fn collapse_binary_branch_fn(self) -> DfResult<Transformed<Self>>

Collapse binary branch with the same operator. I.e., this transformer changes the binary-tree AST into a multiple branching AST.

This function is expected to be called in a bottom-up manner as it wonโt recursion.

fn eliminate_optional_fn(self) -> DfResult<Transformed<Self>>

Eliminate optional pattern. An optional pattern can always be omitted or transformed into a must pattern follows the following rules:

If there is only one pattern and itโs optional, change it to must

If there is any must pattern, remove all other optional patterns

fn eliminate_single_child_fn(self) -> DfResult<Transformed<Self>>

Eliminate single child PatternAst::Binary node. If a binary node has only one child, it can be replaced by its only child.

This function prefers to be applied in a top-down manner. But itโs not required.