common_function::scalars

Module math

Re-exports

pub use clamp::ClampFunction;

pub use clamp::ClampMaxFunction;

pub use clamp::ClampMinFunction;

Modules

clamp

modulo 🔒

pow 🔒

rate 🔒

Structs

MathFunction 🔒

PowFunction

RangeFunction 🔒

RangeFunction will never be used as a normal function, just for datafusion to generate logical plan for RangeSelect

RateFunction

generates rates from a sequence of adjacent data points.