common_function::scalars::vector

Module impl_conv

Functions

as_veclit

Convert a string or binary literal to a vector literal.

as_veclit_if_const

Convert a constant string or binary literal to a vector literal.

binlit_as_veclit

Convert a u8 slice to a vector literal.

parse_veclit_from_strlit

Parse a string literal to a vector literal. Valid inputs are strings like “[1.0, 2.0, 3.0]”.

veclit_to_binlit

Convert a vector literal to a binary literal.