common_function::function

Trait Function

pub trait Function:
    Display
    + Sync
    + Send {
    // Required methods
    fn name(&self) -> &str;
    fn return_type(
        &self,
        input_types: &[ConcreteDataType],
    ) -> Result<ConcreteDataType>;
    fn signature(&self) -> Signature;
    fn eval(
        &self,
        ctx: &FunctionContext,
        columns: &[VectorRef],
    ) -> Result<VectorRef>;
}

Scalar function trait, modified from databend to adapt datafusion TODO(dennis): optimize function by it’s features such as monotonicity etc.

Required Methods

fn name(&self) -> &str

Returns the name of the function, should be unique.

fn return_type( &self, input_types: &[ConcreteDataType], ) -> Result<ConcreteDataType>

The returned data type of function execution.

fn signature(&self) -> Signature

The signature of function.

fn eval( &self, ctx: &FunctionContext, columns: &[VectorRef], ) -> Result<VectorRef>

Evaluate the function, e.g. run/execute the function.