common_function::scalars::geo
Module h3

Structs

H3CellBase

Function that returns base cell of given cell id

H3CellCenterChild

Function that returns center child cell of given cell id

H3CellCenterLatLng

Function that returns centroid latitude and longitude of given cell id

H3CellContains

Tests if cells contains given cells

H3CellDistanceEuclideanDegree

Get Euclidean distance of two cell centroid

H3CellDistanceSphereKm

Get WGS84 great circle distance of two cell centroid

H3CellIsPentagon

Function that check if given cell id is a pentagon

H3CellParent

Function that returns parent cell of given cell id and resolution

H3CellResolution

Function that returns resolution of given cell id

H3CellToChildPos

Function that returns the cell position if its parent at given resolution

H3CellToChildren

Function that returns children cell list

H3CellToChildrenSize

Function that returns children cell count

H3CellToString

Function that converts cell id to its string form

H3ChildPosToCell

Function that returns the cell at given position of the parent at given resolution

H3GridDisk

Function that returns cells with k distances of given cell

H3GridDiskDistances

Function that returns all cells within k distances of given cell

H3GridDistance

Function that returns distance between two cells

H3GridPathCells

Function that returns path cells between two cells

H3LatLngToCell

Function that returns h3 encoding cellid for a given geospatial coordinate.

H3LatLngToCellString

Function that returns h3 encoding cellid in string form for a given geospatial coordinate.

H3StringToCell

Function that converts cell string id to uint64 number

Statics

CELL_TYPES 🔒

COORDINATE_TYPES 🔒

DISTANCE_TYPES 🔒

POSITION_TYPES 🔒

RESOLUTION_TYPES 🔒

Functions

cell_from_value 🔒

cells_from_value 🔒

extract cell array from all possible types including:

signature_of_cell 🔒

signature_of_cell_and_distance 🔒

signature_of_cell_and_resolution 🔒

signature_of_double_cells 🔒

value_to_distance 🔒

value_to_position 🔒

value_to_resolution 🔒