Module common_function::scalars::numpy::interp

Functions

binary_search_ascending_vector 🔒

search the biggest number that smaller than x in xp

concrete_type_to_primitive_vector 🔒

interp

One-dimensional linear interpolation for monotonically increasing sample points. Refers to https://github.com/numpy/numpy/blob/b101756ac02e390d605b2febcded30a1da50cc2c/numpy/core/src/multiarray/compiled_base.c#L491

linear_search_ascending_vector 🔒

search the biggest number that smaller than x in xp